Fetch supported networks' latest blocks at startup
We don't need to subscribe to every supported network, but pulling the
latest block and block height makes it clear to the frontend which
networks are supported by the backend.

Longer term, this might not be necessary as we move to a configurable
network list and custom RPCs... but it's handy for now.